+++ example/ios/RunnerTests/RunnerTests.swift
@@ -10,14 +10,14 @@ import XCTest
 
 class RunnerTests: XCTestCase {
 
-  func testGetPlatformVersion() {
+  func testIsAvailable() {
     let plugin = SignInWithAppleNativePlugin()
 
-    let call = FlutterMethodCall(methodName: "getPlatformVersion", arguments: [])
+    let call = FlutterMethodCall(methodName: "isAvailable", arguments: [])
 
     let resultExpectation = expectation(description: "result block must be called.")
     plugin.handle(call) { result in
-      XCTAssertEqual(result as! String, "iOS " + UIDevice.current.systemVersion)
+      XCTAssertEqual(result as! Bool, true)
       resultExpectation.fulfill()
     }
     waitForExpectations(timeout: 1)