+++ ios/Classes/SignInWithAppleNativePlugin.swift
@@ -8,14 +8,14 @@ public class SignInWithAppleNativePlugin: NSObject, FlutterPlugin {
     let instance = SignInWithAppleNativePlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
 
-    let viewFactory = SignInWithAppleViewFactory.init(messenger: registrar.messenger())
-    registrar.register(viewFactory, withId: "SignInWithAppleNativeButton")
+    if #available(iOS 13, *) {
+        let viewFactory = SignInWithAppleViewFactory.init(messenger: registrar.messenger())
+        registrar.register(viewFactory, withId: "SignInWithAppleNativeButton")
+    }
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
-    case "getPlatformVersion":
-      result("iOS " + UIDevice.current.systemVersion)
     case "isAvailable":
         if #available(iOS 13, *) {
             result(true)