+++ ios/Classes/SignInWithAppleView.swift
@@ -1,14 +1,8 @@
-//
-//  SignInWithAppleView.swift
-//  sign_in_with_apple_native
-//
-//  Created by Władysław Czyżewski on 25.11.2023.
-//
-
 import Foundation
 import Flutter
 import AuthenticationServices
 
+@available(iOS 13, *)
 class SignInWithAppleView : NSObject, FlutterPlatformView, ASAuthorizationControllerDelegate {
     let frame: CGRect
     let viewId: Int64
@@ -22,18 +16,14 @@ class SignInWithAppleView : NSObject, FlutterPlatformView, ASAuthorizationContro
         self.viewId = viewId
         self.args = args
         self.messenger = messenger
-        
+
         self.methodChannel = FlutterMethodChannel(name: "SignInWithAppleNativeButton_\(viewId)", binaryMessenger: messenger)
     }
 
     func view() -> UIView {
-        if #available(iOS 13, *) {
-            let button = ASAuthorizationAppleIDButton(frame: frame)
-            button.addTarget(self, action: #selector(onPressed), for: .touchUpInside)
-            return button
-        }
-
-        return UIView()
+        let button = ASAuthorizationAppleIDButton(frame: frame)
+        button.addTarget(self, action: #selector(onPressed), for: .touchUpInside)
+        return button
     }
 
     @objc func onPressed() {