+++ ios/Classes/SignInWithAppleViewFactory.swift
@@ -1,13 +1,7 @@
-//
-//  SingInWithAppleViewFactory.swift
-//  sign_in_with_apple_native
-//
-//  Created by Władysław Czyżewski on 25.11.2023.
-//
-
 import Foundation
 import Flutter
 
+@available(iOS 13, *)
 class SignInWithAppleViewFactory : NSObject, FlutterPlatformViewFactory {
     let messenger: FlutterBinaryMessenger