+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Changelog
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release
+- Support for iOS 13+ only
+- Support for Apple ID credentials only
+- Native Sign in with Apple button